requirements.txt
pydantic
langchain >= 0.1.13
langchain-community >= 0.0.20
langchain-core
langchain-groq
ipykernel 
python-dotenv 
streamlit
PyPDF2
chromadb
faiss-cpu
google-generativeai
langchain-google-genai
# pdf2image